20240503 수업

+++ src/main/java/com/prj1/controller/BoardController.java
@@ -3,10 +3,12 @@ package com.prj1.controller;
 import com.prj1.domain.Board;
 import com.prj1.service.BoardService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @Controller
@@ -21,10 +23,9 @@ public class BoardController {
     }
 
     @PostMapping("/add")
-    public String addPost(Board board, RedirectAttributes rttr) {
-        System.out.println("board = " + board);
+    public String addPost(Board board, Authentication authentication, RedirectAttributes rttr) {
 
-        service.add(board);
+        service.add(board, authentication);
 
         rttr.addAttribute("id", board.getId());
         return "redirect:/board";
@@ -42,7 +43,8 @@ public class BoardController {
     }
 
     @GetMapping("/")
-    public String home(Integer page, Model model) {
+    public String home(@RequestParam(value = "page", defaultValue = "1") Integer page,
+                       Model model) {
         // 게시물 목록 조회 (select)
         // 모델에 넣고
         model.addAllAttributes(service.list(page));
@@ -51,8 +53,10 @@ public class BoardController {
     }
 
     @PostMapping("delete")
-    public String delete(Integer id) {
-        service.remove(id);
+    public String delete(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+        }
         return "redirect:/";
     }
 
@@ -66,8 +70,10 @@ public class BoardController {
     }
 
     @PostMapping("/modify")
-    public String modifyPost(Board board, RedirectAttributes rttr) {
-        service.modify(board);
+    public String modifyPost(Board board, Authentication authentication, RedirectAttributes rttr) {
+        if (service.hasAccess(board.getId(), authentication)) {
+            service.modify(board);
+        }
         rttr.addAttribute("id", board.getId());
         return "redirect:/board";
     }
